Search Movies and Pagination functionality added
