+++ moviesDeatails/Model/ImageData/DownloadData.swift
@@ -8,18 +8,22 @@
 import Foundation
 import UIKit
 
-
+/// Download Image from Network
 class Downloader{
+    
+    ///Make Class Singleton
     static let shared = Downloader()
     private init(){}
+    
+    ///Cache the Image Data on Local
     class ImageStore: NSObject {
         static let cache = NSCache<NSString, UIImage>()
     }
+    
+    ///Get image
     func getImage(fromUrl urlString: String, completion: ((_ image: UIImage?) -> ())?) {
         if let image = ImageStore.cache.object(forKey: urlString as NSString) {
-            DispatchQueue.main.async {
-                completion?(image)
-            }
+            completion?(image)
         }else{
             guard let url = URL(string: urlString) else {
                 completion?(nil)
@@ -31,7 +35,8 @@ class Downloader{
                     completion?(nil)
                     return
                 }
-                guard let httpResponse = response as? HTTPURLResponse,(200...299).contains(httpResponse.statusCode) else {
+                guard let httpResponse = response as? HTTPURLResponse, httpResponse.statusCode == 200 else {
+                    print("Status code not 200")
                     completion?(nil)
                     return
                 }
