+++ moviesDeatails/Model/JsonData/DataDownload.swift
@@ -10,13 +10,15 @@ import Foundation
 import UIKit
 import Alamofire
 
-
+///Download Api data From Network
 class JsonDownloader{
     
+    ///Make Class Singleton
     static let shared = JsonDownloader()
     private init(){
     }
     
+    /// Get Data Using URLSession
     func apiData(fromUrl urlString: String, completion: ((_ apiData: [MovieData]?, _ totalPages: Int, _ error: String?) -> ())?){
            guard let url = URL(string: urlString) else {
               completion?(nil, 0, urlString)
@@ -36,6 +38,8 @@ class JsonDownloader{
             }
         }.resume()
      }
+    
+    /// Get Data Using AlamoFire
     func apiDataFromAF(fromUrl urlString: String, completion: ((_ apiData: [MovieData]?, _ totalPages: Int, _ error: String?) -> ())?){
         Alamofire.request(urlString, method: .get, parameters: nil, encoding: URLEncoding.default, headers: nil).response{ data in
             guard let data = data.data else {
